requirements.txt
# Working MOM-Bot Requirements
# Only packages that successfully install and work

# Core Dependencies
Cython>=0.29.0

# Web Framework
flask>=3.1.2
werkzeug>=3.1.3
gunicorn>=21.2.0
requests>=2.32.5
pandas>=2.3.3
numpy>=1.23.3

# Configuration Management
omegaconf>=2.3.0
PyYAML>=6.0.3
python-dotenv>=1.1.1

# Audio Processing
librosa>=0.11.0
soundfile>=0.13.1
audioread>=3.0.1
soxr>=1.0.0
pydub>=0.25.1

# Machine Learning Framework
torch>=2.8.0
torchaudio>=2.8.0
torchmetrics>=1.8.2
scikit-learn>=1.3.0
scipy>=1.10.0

# PyTorch Lightning & NeMo Dependencies
lightning>=2.5.5
pytorch-lightning>=2.5.5
hydra-core>=1.3.2
lightning-utilities>=0.15.2

# NeMo Toolkit Dependencies
lhotse>=1.31.1
einops>=0.8.1
nemo-toolkit>=1.20.0

# OpenAI Integration
openai>=2.1.0
openai-whisper>=20240314

# Text Processing
text-unidecode>=1.3
sentencepiece>=0.2.1
transformers>=4.30.0
tokenizers>=0.13.0
huggingface-hub>=0.15.0
safetensors>=0.3.0
regex>=2023.0.0

# Dataset Handling
datasets>=2.10.0
pyarrow>=10.0.0
dill>=0.3.5
multiprocess>=0.70.0
xxhash>=3.0.0

# System Utilities
tqdm>=4.60.0
packaging>=20.0
setuptools>=55.0.0
filelock>=3.0.0
fsspec>=2022.0.0
pooch>=1.6.0

# Monitoring & Logging
tensorboard>=2.13.0

# Configuration & Parsing
antlr4-python3-runtime==4.9.*
onnx>=1.12.0
protobuf>=3.20.0

# Evaluation Metrics
jiwer>=2.3.0
rapidfuzz>=2.0.0
editdistance>=0.5.3

# Pyannote Audio Processing
pyannote.audio>=3.0.1
pyannote-core>=5.0.0
pyannote-metrics>=3.2.0
matplotlib>=3.5.0

# Additional Audio Processing Dependencies
webdataset>=0.2.0
braceexpand>=0.1.5
tabulate>=0.8.9
cytoolz>=0.10.0
intervaltree>=3.0.0
lilcom>=1.8.0
toolz>=0.11.0
sortedcontainers>=2.4.0

# Interactive Python
ipython>=7.30.0
traitlets>=5.8.0
jedi>=0.18.0
matplotlib-inline>=0.1.6

# Core System Dependencies
numba>=0.56.0
llvmlite>=0.39.0
wrapt>=1.13.0
wget>=3.2